Add tests + copyright
