+++ src/mechaphlowers/core/models/balance/functions.py
@@ -1,3 +1,10 @@
+# Copyright (c) 2025, RTE (http://www.rte-france.com)
+# This Source Code Form is subject to the terms of the Mozilla Public
+# License, v. 2.0. If a copy of the MPL was not distributed with this
+# file, You can obtain one at http://mozilla.org/MPL/2.0/.
+# SPDX-License-Identifier: MPL-2.0
+
+
 from functools import lru_cache, wraps
 
 import numpy as np
